Modify Nlp project about the code and TizenFx rule
