+++ src/Tizen.Nlp/Tizen.Nlp.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27703.2035
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NlpCommon.cs", "Tizen.Nlp.csproj", "{AE6ACB66-52B4-409B-9ED4-9BE172EB1309}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tizen.Nlp", "Tizen.Nlp.csproj", "{AE6ACB66-52B4-409B-9ED4-9BE172EB1309}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
